Feature

+++ src/hooks/useProfileLogic.js
@@ -1,4 +1,4 @@
-import { useState, useContext } from 'react';
+import { useState, useContext, useCallback } from 'react';
 import { useFirebaseAuth } from '../hooks';
 import {
   createFileURL,
@@ -8,7 +8,7 @@ import {
 import { getUserDocRef } from '../helpers/manageFirestore';
 import { DataContext, DialogContext, ToastContext } from '../context';
 import { storage } from '../lib/firebase.dev';
-import NoAvatar from "../assets/no-avatar.png"
+import NoAvatar from '../assets/no-avatar.png';
 
 export default function useProfileLogic() {
   const { dataState, dispatchData } = useContext(DataContext);
@@ -19,12 +19,20 @@ export default function useProfileLogic() {
   const [bio, setBio] = useState(dataState?.profile?.bio || '');
   const [errors, setErrors] = useState({});
   const [picture, setPicture] = useState(null);
-  const [pictureURL, setPictureURL] = useState(currentUser?.photoURL || NoAvatar); // !!! Add no picture.
+  const [pictureURL, setPictureURL] = useState(
+    currentUser?.photoURL || NoAvatar
+  ); // !!! Add no picture.
   const [loading, setLoading] = useState(false);
   const [showFileInput, setShowFileInput] = useState(false);
 
+  const isPhotoURLFromGoogle = useCallback(
+    () => /googleusercontent/i.test(currentUser.photoURL),
+    [currentUser.photoURL]
+  );
+
   const handlePictureSubmit = async (e) => {
     e.preventDefault();
+    if (currentUser.photoURL === pictureURL) return;
     setLoading(true);
 
     if (!errors.length && picture) {
@@ -36,7 +44,10 @@ export default function useProfileLogic() {
 
       try {
         if (currentUser.photoURL) {
-          await storage.refFromURL(currentUser.photoURL).delete();
+          // If photoURL is taken from google, it won't try to delete the picture from the firebase storage.
+          if (!isPhotoURLFromGoogle()) {
+            await storage.refFromURL(currentUser.photoURL).delete();
+          }
         }
         await pictureRef.put(picture);
         const photoURL = await pictureRef.getDownloadURL();
@@ -60,7 +71,6 @@ export default function useProfileLogic() {
 
   const handleBioAndNameSubmit = async (e) => {
     e.preventDefault();
-
     const errorObj = {};
 
     if (bio.length > 1200) errorObj.bio = 'Maximum character limit is 1200.';
@@ -100,9 +110,9 @@ export default function useProfileLogic() {
 
   const handleFileChange = (e) => {
     const file = e.target.files[0];
-
+    if (validateFileSize(file, 1)) setPicture(file);
     if (!file) return;
-    setPicture(file);
+
     const errorArr = [];
 
     if (!validateFileSize(file, 1))
@@ -128,12 +138,21 @@ export default function useProfileLogic() {
 
   const handleDeleteClick = (e) => {
     e.stopPropagation();
+
     const deleteUserPicture = async () => {
+      if (!currentUser.photoURL) {
+        return dispatchToast({
+          type: 'ERROR',
+          payload: "You don't have any picture to delete.",
+        });
+      }
       setLoading(true);
       try {
-        await storage.refFromURL(currentUser.photoURL).delete();
+        if (!isPhotoURLFromGoogle()) {
+          await storage.refFromURL(currentUser.photoURL).delete();
+        }
         await updateProfile({ photoURL: '' });
-        setPictureURL('');
+        setPictureURL(NoAvatar);
         dispatchToast({
           type: 'NOTIFICATION',
           payload: 'Picture has been deleted successfully',
